Replacing the HBox full of RadioButtons by the notebook tabs
We spent a lot of code to implement the same behaviour of a tab list of
a GtkNotebook in a GtkHBox full of GtkRadioButton's and there were some
bugs that would be quite hard to fix like implementing scrolling.

For this reason, I'm changing it back to use tabs, at least temporarily. If
we find that notebooks are really ugly (the reason we chosen the other
implementation) we can just implement a widget with the same API and
replace it back.