Fixing a problem when calling `get_fork_params' with the second param
This second param used to receive a dictionary to update default values
that fork params should return. The problem was that now, instead of
creating a dictionary to pass to VteTerminal, we are creating a list, so
it does not make sense to receive a dictionary anymore.

Another point is that there is no sense to change anything but the fork
directory. (till now)