Updating the keybinder library
Now, instead of using a very old version of the keybinder code
a bit home baked, a bit stolen from libegg and a bit from tomboy,
we are now using the libkeybinder.

The problem is that upstream does not provide a gtk3 package,
even knowing that libkeybind is (or seems to be) compatible with
it. This way, instead of linking against it, I had to copy and
paste it again :/